requirements.txt
requests==2.28.1
pandas==2.1.1
matplotlib==3.7.1
python-dotenv==1.0.0
streamlit==1.25.0
ast
smtplib
ssl
email
mimetypes
